feat: simplify
