+++ app/demo3.tsx
@@ -1,9 +1,9 @@
 import { Demo2 } from "@/features/demo2/Demo2";
 import { DemoLayout } from "@/components/DemoLayout";
 
-export default function Demo2Page() {
+export default function Demo3Page() {
   return (
-    <DemoLayout title="Demo 2">
+    <DemoLayout title="Demo 3">
       <Demo2 />
     </DemoLayout>
   );
